requirements.txt
autopep8==1.6.0
brotlipy==0.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cryptography==38.0.1
Deprecated==1.2.13
idna==3.4
mysql-connector-python==8.0.31
pip==22.2.2
protobuf==3.20.1
pycodestyle==2.8.0
pycparser==2.21
PyGithub==1.55
PyJWT==2.4.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
setuptools==65.5.0
six==1.16.0
toml==0.10.2
urllib3==1.26.12
wheel==0.37.1
wrapt==1.14.1
